Move downloads to every streamlit page
