+++ docs/pages/Benchmarks.py
@@ -1,5 +1,7 @@
 import streamlit as st
+from policyengine_us_data.utils.docs_prerequisites import download_data
 
+download_data()
 st.title("Benchmarks")
 
 from policyengine_us_data.datasets import CPS_2024, PUF_2024, EnhancedCPS_2024
@@ -106,7 +108,7 @@ long_to_wide["CPS to ECPS change"] = (
     long_to_wide["Enhanced CPS 2024"] - long_to_wide["CPS 2024 (2022-based)"]
 )
 long_to_wide["PUF to ECPS change"] = (
-    long_to_wide["Enhanced CPS 2024"] - long_to_wide["PUF 2024 (2022-based)"]
+    long_to_wide["Enhanced CPS 2024"] - long_to_wide["PUF 2024 (2015-based)"]
 )
 long_to_wide.sort_values("cps_to_ecps_change", ascending=False)
